app/requirements.txt
#psycopg2
flask
psycopg2-binary
azure-identity
azure-keyvault-secrets